refactor: make checkpoint parsing helpers internal
